Better preparation of results in getAllDevices(). Fixes warning on perl 5.20
Perl 5.20.0 gives a warning on calling each() on a refernce-value, as seen in
lib/Enterasys/NetSight.pm Line 65. So first I chose to separate the "return
undef;" when getting no data. And because the result of the SOAP-Call is an
array of hashes, I chose to use map for building the eesult-hash "%devices"
as this seems to be more appropiate.